[Phase 1] fix: price 타입 int → long 통일
Product.price, OrderLine.price 및 관련 DTO의 타입을 int에서 long으로 변경하여
Order.totalAmount(long)과 타입 일관성을 확보하고, 큰 금액에서의 오버플로우를 방지한다.

+++ src/main/java/com/orderplatform/order/entity/OrderLine.java
@@ -29,12 +29,12 @@ public class OrderLine {
     private String productName;
 
     @Column(nullable = false)
-    private int price;
+    private long price;
 
     @Column(nullable = false)
     private int quantity;
 
-    public OrderLine(Long productId, String productName, int price, int quantity) {
+    public OrderLine(Long productId, String productName, long price, int quantity) {
         this.productId = productId;
         this.productName = productName;
         this.price = price;
